Makes sure that the masterizer, looper and loop recorder all properly update each others navcol to show playing loop. This is necssary in multi screen mode since, you can't wait for the user to change screens to get the update.
I could see in the future, using a subscription event model instead for UI updates between screens, so that a screen just sends an event that something has happened and the other screens can subscribe or not. Mainly for UI related updates. Updates to the data model should be done by direct calls to the underlying model (mode).  

Adam, can you check this change? I tested it but you may know a cleaner way of doing this.