Ok, this is a change that was suggested by the beta feedback and I also agree.
The position bar and key position at the start of recording a pattern is the one that we snap back to at the end of recording. This is more intuitive than what we did in beta 3 which was to leave them in the position you arrived at at the end of the pattern record.