Redesigned 7up clip launcher.
Ok, this was actually harder than I thought it would be.

This component was always expensive in terms of resources because it was polling 7 pages of clips slots for status. Now we don't poll at all. We setup observers on the clips slots.

We already were using them for scenes and for active clip in a track etc, but detecting moving of clips was a real pain.

Mostly working as I want but a few small niggles like moving a clip that is playing while it is playing leaves the old slot active.

Added a reset button just incase, so that users can do a manual poll if it gets out of sync.